* develop:
  Add version 0.0.1
  Fix ReserveBook endpoint
  Refactor LoanServiceTest
  Fix LoanService implementation
  Fix base rest exception
  Fix tests covering acceptance criteria
  Add first feature implementation
  Add tests covering acceptance criteria
  Add feature skeleton

+++ LibraryLoans.Core/Exceptions/MalformedRestException.cs
@@ -5,5 +5,5 @@ namespace LibraryLoans.Core.Exceptions;
 
 public class MalformedRestException : BaseRestException
 {
-    public MalformedRestException(string message) : base((int) HttpStatusCode.BadRequest, message) { }
+    public MalformedRestException(string message) : base(HttpStatusCode.BadRequest, message) { }
 }
